refactor: clear prev project
