+++ package.json
@@ -4,7 +4,8 @@
   "version": "0.0.0",
   "type": "module",
   "scripts": {
-    "dev": "next dev",
+    "dev": "node ./server.js",
+    "start": "cross-env NODE_ENV=production node ./server.js",
     "build": "next build",
     "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
     "format:fix": "npx prettier src --write",
@@ -15,7 +16,11 @@
   },
   "dependencies": {
     "@reduxjs/toolkit": "^2.2.7",
+    "@remix-run/node": "^2.11.1",
+    "@remix-run/react": "^2.11.1",
+    "@remix-run/serve": "^2.11.1",
     "classnames": "^2.5.1",
+    "isbot": "^4.4.0",
     "msw": "^2.3.4",
     "next": "^14.2.5",
     "react": "^18.3.1",
@@ -28,6 +33,7 @@
     "react-router-dom": "^6.25.1"
   },
   "devDependencies": {
+    "@remix-run/dev": "^2.11.1",
     "@testing-library/dom": "^10.4.0",
     "@testing-library/jest-dom": "^6.4.8",
     "@testing-library/react": "^16.0.0",
@@ -40,7 +46,7 @@
     "@typescript-eslint/parser": "^7.15.0",
     "@vitejs/plugin-react": "^4.3.1",
     "@vitest/coverage-v8": "^2.0.3",
-    "autoprefixer": "^10.4.19",
+    "autoprefixer": "^10.4.20",
     "eslint": "^8.57.0",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-config-next": "^14.2.5",
@@ -58,14 +64,14 @@
     "jest-transformer-svg": "^2.0.2",
     "jsdom": "^24.1.1",
     "lint-staged": "^15.2.7",
-    "postcss": "^8.4.39",
+    "postcss": "^8.4.41",
     "prettier": "3.3.2",
-    "tailwindcss": "^3.4.6",
+    "tailwindcss": "^3.4.10",
     "ts-jest": "^29.2.4",
     "ts-node": "^10.9.2",
     "typescript": "^5.2.2",
-    "vitest": "^2.0.3",
-    "vite": "5.4.0"
+    "vite": "^5.4.0",
+    "vitest": "^2.0.3"
   },
   "lint-staged": {
     "**/*.{ts,tsx}": [
