--- src/app/components/App.tsx
@@ -1,23 +1,0 @@
-'use client';
-
-import { createContext, useState } from 'react';
-
-export const ThemeContext = createContext({
-  theme: 'light',
-  toggleTheme: () => {},
-});
-
-const App = () => {
-  const [theme, setTheme] = useState<string>('light');
-  const toggleTheme = () => {
-    theme === 'light' ? setTheme('dark') : setTheme('light');
-  };
-
-  return (
-    <ThemeContext.Provider
-      value={{ theme: theme, toggleTheme: toggleTheme }}
-    ></ThemeContext.Provider>
-  );
-};
-
-export default App;
