--- src/app/index.css
@@ -1,13 +1,0 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
-body {
-  margin: 0;
-  height: 100vh;
-}
-
-p {
-  margin: 0;
-  padding: 0;
-}
