--- src/app/page.tsx
@@ -1,5 +1,0 @@
-import { permanentRedirect } from 'next/navigation';
-
-export default function IndexPage() {
-  permanentRedirect(`/page/1`);
-}
