--- src/app/page/ClientOnly.tsx
@@ -1,10 +1,0 @@
-'use client';
-
-import dynamic from 'next/dynamic';
-import React from 'react';
-
-const App = dynamic(() => import('../components/App'), { ssr: false });
-
-export function ClientOnly() {
-  return <App />;
-}
