--- src/app/page/page.tsx
@@ -1,13 +1,0 @@
-import { Provider } from 'react-redux';
-import { ClientOnly } from './ClientOnly';
-
-import React from 'react';
-import { store } from '../store/store';
-
-export default function Page() {
-  return (
-    <Provider store={store}>
-      <ClientOnly />
-    </Provider>
-  );
-}
