+++ src/tests/MainPage.test.tsx
@@ -1,8 +1,8 @@
-import { Item } from '../app/components/Item';
-import { ItemList } from '../app/components/ItemList';
-import { Search } from '../app/components/Search';
-import { MainPage } from '../app/myPages/MainPage';
-import { store } from '../app/store/store';
+import { Item } from '../../app/components/Item';
+import { ItemList } from '../../app/components/ItemList';
+import { Search } from '../../app/components/Search';
+import { MainPage } from '../../app/myPages/MainPage';
+import { store } from '../../app/store/store';
 import { render } from '@testing-library/react';
 import { Provider } from 'react-redux';
