+++ src/tests/MyModal.test.tsx
@@ -1,5 +1,5 @@
-import { MyModal } from '../app/components/MyModal';
-import { store } from '../app/store/store';
+import { MyModal } from '../../app/components/MyModal';
+import { store } from '../../app/store/store';
 import { render, screen } from '@testing-library/react';
 import { Provider } from 'react-redux';
