+++ test-output.json
@@ -1,1 +1,1 @@
-{"numTotalTestSuites":5,"numPassedTestSuites":5,"numFailedTestSuites":0,"numPendingTestSuites":0,"numTotalTests":11,"numPassedTests":11,"numFailedTests":0,"numPendingTests":0,"numTodoTests":0,"snapshot":{"added":0,"failure":false,"filesAdded":0,"filesRemoved":0,"filesRemovedList":[],"filesUnmatched":0,"filesUpdated":0,"matched":0,"total":0,"unchecked":0,"uncheckedKeysByFile":[],"unmatched":0,"updated":0,"didUpdate":false},"startTime":1723724530257,"success":true,"testResults":[{"assertionResults":[{"ancestorTitles":[],"fullName":"should render button","status":"passed","title":"should render button","duration":87,"failureMessages":[],"meta":{}},{"ancestorTitles":[],"fullName":"should render ThemeContext","status":"passed","title":"should render ThemeContext","duration":22,"failureMessages":[],"meta":{}},{"ancestorTitles":[],"fullName":"should contain correct url","status":"passed","title":"should contain correct url","duration":3,"failureMessages":[],"meta":{}},{"ancestorTitles":[],"fullName":"should render button","status":"passed","title":"should render button","duration":163,"failureMessages":[],"meta":{}},{"ancestorTitles":[],"fullName":"should render button","status":"passed","title":"should render button","duration":31,"failureMessages":[],"meta":{}}],"startTime":1723724532461,"endTime":1723724532768,"status":"passed","message":"","name":"C:/Users/nikit/Desktop/Курсы/RsSchool(Stage3)/REACT2024Q3/src/tests/App.test.tsx"},{"assertionResults":[{"ancestorTitles":[],"fullName":"renders a message","status":"passed","title":"renders a message","duration":106,"failureMessages":[],"meta":{}},{"ancestorTitles":[],"fullName":"renders a message","status":"passed","title":"renders a message","duration":134,"failureMessages":[],"meta":{}}],"startTime":1723724532391,"endTime":1723724532631,"status":"passed","message":"","name":"C:/Users/nikit/Desktop/Курсы/RsSchool(Stage3)/REACT2024Q3/src/tests/ErrorBoundary.test.tsx"},{"assertionResults":[{"ancestorTitles":[],"fullName":"renders a message","status":"passed","title":"renders a message","duration":56,"failureMessages":[],"meta":{}}],"startTime":1723724532283,"endTime":1723724532339,"status":"passed","message":"","name":"C:/Users/nikit/Desktop/Курсы/RsSchool(Stage3)/REACT2024Q3/src/tests/ErrorPage.test.tsx"},{"assertionResults":[{"ancestorTitles":[],"fullName":"renders a Search component","status":"passed","title":"renders a Search component","duration":187,"failureMessages":[],"meta":{}},{"ancestorTitles":[],"fullName":"renders a ItemList component","status":"passed","title":"renders a ItemList component","duration":9,"failureMessages":[],"meta":{}}],"startTime":1723724532488,"endTime":1723724532684,"status":"passed","message":"","name":"C:/Users/nikit/Desktop/Курсы/RsSchool(Stage3)/REACT2024Q3/src/tests/MainPage.test.tsx"},{"assertionResults":[{"ancestorTitles":[],"fullName":"renders a message","status":"passed","title":"renders a message","duration":220,"failureMessages":[],"meta":{}}],"startTime":1723724532379,"endTime":1723724532599,"status":"passed","message":"","name":"C:/Users/nikit/Desktop/Курсы/RsSchool(Stage3)/REACT2024Q3/src/tests/MyModal.test.tsx"}]}
+{"numTotalTestSuites":5,"numPassedTestSuites":5,"numFailedTestSuites":0,"numPendingTestSuites":0,"numTotalTests":11,"numPassedTests":11,"numFailedTests":0,"numPendingTests":0,"numTodoTests":0,"snapshot":{"added":0,"failure":false,"filesAdded":0,"filesRemoved":0,"filesRemovedList":[],"filesUnmatched":0,"filesUpdated":0,"matched":0,"total":0,"unchecked":0,"uncheckedKeysByFile":[],"unmatched":0,"updated":0,"didUpdate":false},"startTime":1723724617644,"success":true,"testResults":[{"assertionResults":[{"ancestorTitles":[],"fullName":"should render button","status":"passed","title":"should render button","duration":92,"failureMessages":[],"meta":{}},{"ancestorTitles":[],"fullName":"should render ThemeContext","status":"passed","title":"should render ThemeContext","duration":18,"failureMessages":[],"meta":{}},{"ancestorTitles":[],"fullName":"should contain correct url","status":"passed","title":"should contain correct url","duration":3,"failureMessages":[],"meta":{}},{"ancestorTitles":[],"fullName":"should render button","status":"passed","title":"should render button","duration":162,"failureMessages":[],"meta":{}},{"ancestorTitles":[],"fullName":"should render button","status":"passed","title":"should render button","duration":43,"failureMessages":[],"meta":{}}],"startTime":1723724619898,"endTime":1723724620217,"status":"passed","message":"","name":"C:/Users/nikit/Desktop/Курсы/RsSchool(Stage3)/REACT2024Q3/src/tests/App.test.tsx"},{"assertionResults":[{"ancestorTitles":[],"fullName":"renders a message","status":"passed","title":"renders a message","duration":114,"failureMessages":[],"meta":{}},{"ancestorTitles":[],"fullName":"renders a message","status":"passed","title":"renders a message","duration":131,"failureMessages":[],"meta":{}}],"startTime":1723724619827,"endTime":1723724620072,"status":"passed","message":"","name":"C:/Users/nikit/Desktop/Курсы/RsSchool(Stage3)/REACT2024Q3/src/tests/ErrorBoundary.test.tsx"},{"assertionResults":[{"ancestorTitles":[],"fullName":"renders a message","status":"passed","title":"renders a message","duration":47,"failureMessages":[],"meta":{}}],"startTime":1723724619723,"endTime":1723724619770,"status":"passed","message":"","name":"C:/Users/nikit/Desktop/Курсы/RsSchool(Stage3)/REACT2024Q3/src/tests/ErrorPage.test.tsx"},{"assertionResults":[{"ancestorTitles":[],"fullName":"renders a Search component","status":"passed","title":"renders a Search component","duration":158,"failureMessages":[],"meta":{}},{"ancestorTitles":[],"fullName":"renders a ItemList component","status":"passed","title":"renders a ItemList component","duration":11,"failureMessages":[],"meta":{}}],"startTime":1723724619891,"endTime":1723724620060,"status":"passed","message":"","name":"C:/Users/nikit/Desktop/Курсы/RsSchool(Stage3)/REACT2024Q3/src/tests/MainPage.test.tsx"},{"assertionResults":[{"ancestorTitles":[],"fullName":"renders a message","status":"passed","title":"renders a message","duration":198,"failureMessages":[],"meta":{}}],"startTime":1723724619810,"endTime":1723724620008,"status":"passed","message":"","name":"C:/Users/nikit/Desktop/Курсы/RsSchool(Stage3)/REACT2024Q3/src/tests/MyModal.test.tsx"}]}
